% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_analysis.R
\name{Gro_analysis}
\alias{Gro_analysis}
\title{Flag outliers in individual trajectories}
\usage{
Gro_analysis(data_weight, variable_id = "anonID")
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the following columns : \emph{MeasurementValue} (\code{numeric}), \emph{Age}  (\code{numeric})}

\item{variable_id}{\code{character} name of the variable including individual ids. Defaut = "anonID". It must also be a column of data_weight}
}
\value{
data frame; the data set including an additional column \emph{Keep2} of 0 (potential outliers) and 1
}
\description{
This function flag
}
\examples{
data(weights)

}
