% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_cleanmeasures.R
\name{Gro_cleanmeasures}
\alias{Gro_cleanmeasures}
\title{Prepare weight and length Dataset}
\usage{
Gro_cleanmeasures(
  data,
  coresubse,
  CaptiveBirths = TRUE,
  type = "weight",
  MeasureType = NULL,
  InclUnkSex = FALSE,
  mindate = "1980-01-01",
  corevariable = "",
  variablekeep = ""
)
}
\arguments{
\item{data}{\code{data.frame} including the following columns \emph{MeasurementValue}, \emph{MeasurementDate}, \emph{CollectionScopeType}, \emph{UnitOfMeasure}, \emph{ExcludedFromNorms}, \emph{EstimatedMeasurement}, \emph{RecordType}, \emph{MeasurementType}, \emph{anonID} and \emph{AnonInstitutionID}.}

\item{coresubse}{\code{data.frame} including at least the following columns \emph{binSpecies}, \emph{Sex}, \emph{anonID} and \emph{Birthdate} (\code{date})}

\item{CaptiveBirths}{\code{logical} Should captive born individuals be included, only? Default =  TRUE}

\item{type}{\code{character} Either 'weight' or 'length'. Default =  'weight}

\item{MeasureType}{\code{vector of characters} Name of the type of measurements that should be included. Default = NULL, all measurement type are included.}

\item{InclUnkSex}{\code{logical} Should undetermined sexes be included Default =  FALSE}

\item{mindate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{corevariable}{\code{vector character} Name of the core columns that must be added to data}

\item{variablekeep}{\code{vector character} Name of the data columns that must also be returned}
}
\value{
The data frame including selected measures, plus individual birth date and individual age at each measure
}
\description{
Select clean measures, unify the unity used and add individual age at each measure
}
\examples{
data(raw_weights)
data(core)

data_weights= Gro_cleanmeasures(raw_weights, core,
                                CaptiveBirths = TRUE,
                                MeasureType = 'Live weight',
                                InclUnkSex = FALSE, 
                                mindate = "1980-01-01")
}
