% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_lin_ind.R
\name{Gro_lin_ind}
\alias{Gro_lin_ind}
\title{Flag outliers in individual trajectories}
\usage{
Gro_lin_ind(
  data_weight,
  perc_weight_min = 0.2,
  perc_weight_max = 2.5,
  IQR = 1.5,
  remove_ext = T,
  traj_ind = T
)
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the following columns : \emph{MeasurementValue} (\code{numeric}), \emph{Age}  (\code{numeric}) and \emph{anonID}}

\item{perc_weight_min}{\code{numeric} Minimum percentage of weight that an individual can naturally lose or gain during a year. Default =  0.2 (20\%)}

\item{perc_weight_max}{\code{numeric} Maximum percentage of weight that an individual can naturally lose or gain during a year. Default =  2.5 (250\%). This value is used only after 1 year old}

\item{IQR}{\code{numeric} influences the sensitivity of the function to remove outliers. It should be above 1 Default =  1.5. A higher number makes the function less sensitive to find outliers.}

\item{remove_ext}{\code{logical} Do you want to keep the two first and two last observations of individual trajectories? These points may often be highlighted as false outliers. Default =  TRUE}

\item{traj_ind}{\code{logical} Do you want to build linear model at the level of the individual trajectory  Default =  TRUE. If yes, each individual trajectory must have at least 5 measures.}
}
\value{
data frame; the data set including an additional column \emph{Keep2} of 0 (potential outliers) and 1
}
\description{
This function flag
}
\details{
The function uses a log-linear model of the measurement value in relation to a quadratic effect of age for each individual trajectories. It flags as outliers all measures for which the residuals from the linear model are above  IQR * the inter quantile interval (quantile(0.9) - quantile(0.1)) of the residuals. \code{perc_weight_min} and \code{perc_weight_max} limit the values that must be considered as outliers.
The \code{remove_ext} argument prevents the function from removing the 2 last and 2 first value observations in relation to age as these values may often be farther from the prediction line.
}
\examples{
data(weights)

#We use only trajectories for which the number of individual datapoints is > 6
weights <- weights\%>\%
  dplyr::group_by(anonID)\%>\%
  dplyr::mutate(nb= dplyr::n())\%>\%
  dplyr::ungroup()\%>\%
  dplyr::filter(nb >6)

weights_ind <- weights\%>\%Gro_lin_ind(traj_ind = TRUE)
as.numeric(weights_ind$Keep2)

weights_pop <- weights\%>\%Gro_lin_ind(traj_ind = FALSE)
}
