% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_modsettings.R
\name{Gro_ModSettings}
\alias{Gro_ModSettings}
\title{Growth model settings}
\usage{
Gro_ModSettings(data, model = "vonBertalanffy")
}
\arguments{
\item{data}{\code{data.frame} including at least the numeric columns \emph{logx} and \emph{logz}}

\item{model}{\code{character} Name of the model to fit. The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial}
}
\value{
a list including the function of the growth model,  the number of parameter, their names, their initial values and their lowest possible values
}
\description{
This function gets the parameter and equations for the growth model
}
\examples{
logx <- rnorm(100, 0, 1)
logz <- 0.2+ 15 * (1 - exp(-(1) * logx)) +rnorm(100, 0, 0.01)
dat = data.frame(logx = logx, logz = logz)

Gro_ModSettings(data = dat, model = "vonBertalanffy")
}
