% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_outplot.R
\name{Gro_outplot}
\alias{Gro_outplot}
\title{Plot outliers}
\usage{
Gro_outplot(data, title = "", ylimit = NULL, xlimit = NULL)
}
\arguments{
\item{data}{\code{data.frame} including at least the numeric columns \emph{Age}, \emph{MeasurementValue}, \emph{keep1}, \emph{keep2}, \emph{keep3}, and \emph{KEEP}}

\item{title}{\code{character} Title of the plot}

\item{ylimit}{\code{vector of 2 numeric} y limit for the plot}

\item{xlimit}{\code{vector of 2 numeric} x limit for the plot}
}
\value{
Plot of the data points with outliers showed:
\itemize{
\item in red if detected using the maximum possible value
\item in green if detected using the inter quartile range
\item in cyan if detected using individual trajectories
\item in blue if detected from the global log-linear model
}
}
\description{
Plot higlighting the outliers found by \code{Gro_remoutliers()}
}
\examples{
data(weights)
weights = Gro_remoutliers(weights, taxa = "Reptilia", ageMat = 10)

p <- Gro_outplot(data = weights, 
                 title = "Species A", ylimit = c(0,2000))
p
}
