% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_plot.R
\name{Gro_plot}
\alias{Gro_plot}
\title{Plot predicted growth curve}
\usage{
Gro_plot(data, data_percent, title = "")
}
\arguments{
\item{data}{\code{data.frame} including at least the numeric columns \emph{Age}, \emph{MeasurementValue} and \emph{AnimalAnonID}}

\item{data_percent}{\code{data.frame} including the predicted percentile from the growth model and  the numeric columns \emph{Age}, \emph{percentiles50} and other percentiles to plot.}

\item{title}{\code{character} Title of the plot}
}
\value{
Plot
}
\description{
Plot predicted percentiles from a growth model with data points
}
\examples{
Age <- sample(c(0:10), 100, replace = TRUE)
AnimalAnonID <- sample(c(0:20), 100, replace = TRUE)
MeasurementValue <- exp(0.2+15 * (1 - exp(-(0.1) * log(Age+1)))+ 
                          rnorm(100,0,0.01) + AnimalAnonID*0.1)-1 
dat = data.frame(Age = Age, MeasurementValue = MeasurementValue, AnimalAnonID = AnimalAnonID)

out = Gro_analysis(dat, 
                   all_mods = c("logistic", "vonBertalanffy", "gam"), 
                   percentiles = c(2.5, 97.5))

p <- Gro_plot(data = dat, 
              data_percent = out$percent, 
              title =out$AIC_tab$model[1])
p
}
