% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_remoutliers.R
\name{Gro_remoutliers}
\alias{Gro_remoutliers}
\title{Remove outliers from weigth or length data}
\usage{
Gro_remoutliers(
  data_weight,
  taxa,
  ageMat = 0,
  maxweight = NULL,
  variableid = "AnonID",
  min_Nmeasures = 7,
  perc_weight_min = 0.2,
  perc_weight_max = 2.5,
  IQR = 2.75,
  minq = 0.025,
  Ninterval_juv = 500
)
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the following columns : \emph{MeasurementValue} (\code{numeric}), \emph{Age}  (\code{numeric})}

\item{taxa}{\code{character} the name of the taxa studied}

\item{ageMat}{\code{numeric} the age at sexual maturity to differentiate juveniles (still growing) from adults}

\item{maxweight}{\code{numeric} the maximum weight allowed for the dataset}

\item{variableid}{\code{character} name of the variable including individual ids. Defaut = "AnonID". It must also be a column of data_weight.Default = 6. It must be at least 5}

\item{min_Nmeasures}{\code{integer} Minimum number of measures for an individual to check outliers along its growth trajectory. Default = 7. It must be at least 5}

\item{perc_weight_min}{\code{numeric} Minimum percentage of weight that an individual can naturally lose or gain during a year . Default =  0.2 (20\%)}

\item{perc_weight_max}{\code{numeric} Maximum percentage of weight that an individual can naturally lose or gain during a year . Default =  2.5 (250\%)}

\item{IQR}{\code{numeric} influences the sensitivity of the function to remove outliers using log-linear model of growth. It should be above 1 Default =  2.75. A higher number makes the function less sensitive to find outliers.}

\item{minq}{\code{numeric} Sensitivity of the function to remove outliers using percentiles, between 0 and 1. Default =  0.025}

\item{Ninterval_juv}{\code{integer} Minimum number of juveniles used to make one interval of the sliding windows. Default = 500}
}
\value{
The data frame including the additional column \code{KEEP} a numeric vector of 1 and 0 indicating the measures to keep. The 0 signal outliers. Other additional columns keep1, keep2, keep3 indicates the individuals highlighted as outliers (0) in steps 1 to 3.
}
\description{
Takes a data frame including weight measures and Age and look for possible outliers
}
\details{
This function follow 4 different steps to highlight outliers:
1/ It removes the measures higher than \code{maxweight} (Instead of using the OrdMag Fernando that was removing only very extreme values, a maximum value was set for each taxa).
2/ Independently on adults and juveniles, it uses the function Gro_Rout_quan() to removes outliers based on percentiles on adults and juveniles. It uses sliding windows of age for juveniles.
3/  Independently on adults and juveniles, it uses the function Gro_lin_ind() to build generalized additive models for each individual trajectories with at least 7 measures and remove outliers based on the residuals of the models.
4/  It uses the function Gro_lin_ind() to build a common generalized additive model for all growth trajectories and remove outliers based on the residuals of the models.
}
\examples{
data(weights)
weights = Gro_remoutliers(weights, taxa = "Mammalia", ageMat = 10)

plot(MeasurementValue ~Age, data = weights)
#Show the four steps used to remove outliers
#1/ using maximum value
points(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(keep1==0), col = "red", pch = 16) 
#2/ using Inter quartile range
points(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(keep2==0, keep1==1 ), col = "green", pch = 16) 
#3a/ using individual trajectories
points(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(keep3==0, keep2==1), col = "cyan", pch = 16) 
#3b/ using global log-linear model
points(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(KEEP==0, keep3==1), col = "blue", pch = 16) 

#Checking outliers highlighted in Individual trajectories
anon = unique(weights$AnonID[weights$keep3==0 & weights$keep2==1 ])
plot(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(keep2==1, AnonID == anon))
points(MeasurementValue ~Age, data = weights\%>\%dplyr::filter(keep3==0, AnonID == anon, keep2==1), col = "cyan", pch = 16)
}
