% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_rout_quan.R
\name{Gro_Rout_quan}
\alias{Gro_Rout_quan}
\title{Remove outliers from percentiles}
\usage{
Gro_Rout_quan(z, x, minq = 0.05, type = "both")
}
\arguments{
\item{z}{\code{numeric vector} Measurement values}

\item{x}{\code{numeric vector} Age}

\item{minq}{\code{numeric} Sensitivity of the function to remove outliers, between 0 and 1. Default =  0.05}

\item{type}{\code{character} "both", "lower", or "upper" to choose if the lower or upper part of the distribution should be screened. Default =  0.05}
}
\value{
Vector of 0 and 1 indicating the measurement values to keep (=1) vs. the outliers (=0).
}
\description{
Remove outliers from percentiles
}
\details{
This functions removes all values lower and higher than the median value multiplied by how many times the (100-q)th percentile is higher than the qth percentile of the distribution of weights.
}
\examples{
z = c(rnorm(100,5,1), runif(3,40,100))
x = rnorm(103,3,2)
Gro_Rout_quan (z,x, minq = 0.05)
}
