% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_age.R
\name{KM_age}
\alias{KM_age}
\title{Age from Kaplan Meier}
\usage{
KM_age(KM_tab, Lx)
}
\arguments{
\item{KM_tab}{\code{data.frame} The kaplan-Meier table including the columns: \emph{Ages} and \emph{ple}}

\item{Lx}{\code{vector of numeric} Values of Lx for which age is requested}
}
\value{
a data frame including Lx and Age
}
\description{
Estimate Ages at which Lx equals given values from the Kaplan-Meier table. Ages are linearly extrapolated.
}
\examples{
KM_tab = data.frame( Ages = 1:10,
                     ple = sort(runif(10, 0, 1), decreasing = TRUE))
out <- KM_age(KM_tab, Lx = c(0.5,0.1))
}
