% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_lx.R
\name{KM_Lx}
\alias{KM_Lx}
\title{Lx from Kaplan Meier}
\usage{
KM_Lx(KM_tab, Age)
}
\arguments{
\item{KM_tab}{\code{data.frame} The kaplan-Meier table including the columns: \emph{Ages} and \emph{ple}}

\item{Age}{\code{vector of numeric} Ages for which Lx is requested}
}
\value{
a data frame including age, the mean and 95\% credible interval of the L90
}
\description{
Estimate Lx  at given ages from the Kaplan-Meier table.
}
\examples{
KM_tab = data.frame( Ages = 1:10,
                     ple = sort(runif(10, 0, 1), decreasing = TRUE))
out <- KM_Lx(KM_tab, Age = c(3,6.5))
}
