% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zimsdata.R
\name{Load_Zimsdata}
\alias{Load_Zimsdata}
\title{Load Zims data}
\usage{
Load_Zimsdata(taxa, ZIMSdir, extractDate, type = "core", silent = FALSE)
}
\arguments{
\item{taxa}{\code{character} the name of the taxa studied}

\item{ZIMSdir}{\code{character} directory where to find data}

\item{extractDate}{\code{character 'YYYY-MM-DD'} Date of data extraction}

\item{type}{\code{vector character} the type of data to load. Default = 'core'}

\item{silent}{\code{logical} Whether information of advancement should be printed}
}
\value{
A list of the data frames
}
\description{
Check if selected data are already in the global environment. If not, it load them.
}
\details{
\code{type} can take the following values : "core", "health", "institution", "moves", "parent", "weights", "lengths", "collections", "simplified_collections"
}
\examples{

# #example of an environmental ncdf file saved with the package
# file = system.file("coretest.csv", package = 'ISRverse')
# ZIMSdirtest = dirname(file)
# 
# data <- Load_Zimsdata	(taxa = "Mammalia", 
#                            ZIMSdir = ZIMSdirtest, 
#                            extractDate = "2023-12-04", 
#                            type = 'core') 
# data$core
# unlink(ZIMSdirtest, recursive = TRUE)

}
