% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary2.R
\name{make_summary2}
\alias{make_summary2}
\title{Summarize taxon profiles analyses}
\usage{
make_summary2(
  AnalysisDir,
  SaveDir,
  namefile = "",
  TaxaList = "Mammalia",
  BySex = list(Mammalia = c("Male", "Female")),
  Sections = c("sur"),
  ModelSur = "from0"
)
}
\arguments{
\item{AnalysisDir}{\code{character} Directory where to find the analyses results: .Rdata files.}

\item{SaveDir}{\code{character} Directory where to save summary plots and tables.}

\item{namefile}{\code{character} Suffix to add to the name of files produced if needed.}

\item{TaxaList}{\code{vector of character} names of the taxa studied.}

\item{BySex}{\code{list} of the taxa names indicating the sexes analyzed.}

\item{Sections}{\code{vector of character} names of the sections: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")}

\item{ModelSur}{\code{character} Name of the survival models to save}
}
\value{
It saves :
\itemize{
\item A general summary table including:
\itemize{
\item NRaw : Raw number of individuals
\item NDate : Number of individuals with an entry date posterior to the minimum date
\item NGlobal: Number of individuals selected from global Collections
\item NBirthType: Number of individuals selected from birth type
\item NUncertBirth: Number of individuals selected from uncertainty in birth
\item NAlive: Number of individuals still alive
\item FirstDate: Date of first record
\item MaxAgeRaw: Maximum observed age.
}
\item A table AnyAna including:
\itemize{
\item Nselect: the number of individuals of this sex
\item Nlifespan : the number of individuals with estimated lifespan (i.e. estimated birth dates, censored individuals are also included)
\item GapThresh : The threshold value selected for the distribution of longevity
\item NThres : the number of individuals selected using this threshold
\item Surv_Ana: whether the survival analysis has been run
\item Surv_error: if no survival analyses: the error explaining why
\item Surv_model: if the survival analysis run, the name of the model selected
}
\item A table for the survival section including main metrics and checks of the survival analysis. See ?Sur_main() for more details
}
}
\description{
Produce summary tables of the demographic analyses made for all species
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
AnalysisDir  = paste0(dirname(file),'/Rdata')
SaveDir = paste0(tempdir(check = TRUE),'/temp')
dir.create(SaveDir)

SummTab <- make_summary2(AnalysisDir, SaveDir,
                         TaxaList = "Reptilia",
                         BySex = list(Reptilia = c("Male", "Female")) ,
                         Sections = c("sur")
)
list.files(SaveDir)
unlink(SaveDir, recursive = TRUE)
}
