% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surfigtab.R
\name{Make_SurFigTab}
\alias{Make_SurFigTab}
\title{Data Papers: Make figures & Tables}
\usage{
Make_SurFigTab(
  AnalysisDir,
  SaveDir = NA,
  namefile = "",
  Sections = c("sur", "rep", "gro")
)
}
\arguments{
\item{AnalysisDir}{\code{character} directory where to find the analyses results: .Rdata files}

\item{SaveDir}{\code{character} directory where to save summary plots and tables}

\item{namefile}{\code{character} Suffix to add to the name of files produced if needed.}

\item{Sections}{\code{vector of character} names of the sections: "sur", "rep" and/or "gro".}
}
\value{
It saves the plots and tables used in the data papers.
}
\description{
Data Papers: Make figures & Tables
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
AnalysisDir  = dirname(file)
SaveDir = paste0(tempdir(check = TRUE),'/temp')
dir.create(SaveDir)
out <-Make_SurFigTab (AnalysisDir, SaveDir,
                      Sections = "sur")
unlink(SaveDir, recursive = TRUE)
}
