% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\docType{data}
\name{moves}
\alias{moves}
\title{Moves data}
\format{
A data.frame object with 1851 rows and 5 variables:
\describe{
\item{AnimalAnonID}{Anonimized ID of the Animal}
\item{EventType}{Name of the event}
\item{Date}{Date of the move/transaction/event}
\item{From}{Anonimized ID of institution of departure}
\item{To}{Anonimized ID of institution of arrival}
}
}
\source{
Subset of a toy database created by the Species360 Development team in 08-2024
}
\usage{
data(moves)
}
\description{
A simulated dataset including information about individual moves between institutions.
}
\keyword{datasets}
