% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_main.R
\name{Rep_main}
\alias{Rep_main}
\title{Main Reproduction analysis for taxon profiles}
\usage{
Rep_main(
  coresubset,
  collection,
  parent,
  move,
  Repsect = c("agemat", "litter"),
  BirthType_parent = "Captive",
  BirthType_offspring = "Captive",
  Global = TRUE,
  minInstitution = 2,
  minNrepro = 100,
  minNparepro = 30,
  parentProb = 80,
  minNlitter = 20,
  Nday = 7,
  minNseas = 50
)
}
\arguments{
\item{coresubset}{\code{data frame} cleaned core data including only the selected individuals. It must includes at least the following columns: \emph{AnimalAnonID}, 'BirthType', \emph{MaxBirthDate}, \emph{MinBirthDate}, \emph{DepartDate}, \emph{BirthDate},\emph{SexType}, and \emph{FirstHoldingInstitution}}

\item{collection}{\code{data frame} Collection data including at least the following columns: \emph{AnimalAnonID}, \emph{ScopeType}, \emph{ChangeDate}}

\item{parent}{\code{data frame} Parent data including at least the following columns: \emph{ParentAnonID}, \emph{ParentCollectionScopeType}, \emph{OffspringCollectionScopeType}, \emph{AnimalAnonID}, \emph{ParentOriginType}, \emph{Probability}}

\item{move}{\code{data frame} Moves data including at least the following columns: \emph{AnimalAnonID}, \emph{To}, \emph{Date}}

\item{Repsect}{\code{character} names of the reproductive analyses to run: "agemat", "litter" and/or ...}

\item{BirthType_parent}{\code{character} Captive, Wild, or All. Default =  "Captive"}

\item{BirthType_offspring}{\code{character} Captive, Wild, or All. Default =  "Captive"}

\item{Global}{\code{logical} Whether only individuals belonging to global collections should be used. Default = #'}

\item{minInstitution}{\code{numeric} Minimum number of institutions that should hold records to run the survival analysis. Default = 2}

\item{minNrepro}{\code{numeric} Minimum number of birth records needed to run reproductive analyses. Default = 50}

\item{minNparepro}{\code{numeric} Minimum number of unique parent records needed to run reproductive analyses. Default = 30}

\item{parentProb}{\code{numeric} Minimum percentage of parentage probability to include. Default = 80}

\item{minNlitter}{\code{numeric} Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases. Default = 30}

\item{Nday}{\code{numeric} Number of consecutive days over which the birth dates of a litter/clutch can be spread. Default = 7}

\item{minNseas}{\code{numeric} XXXXXXXXXXX}
}
\value{
The output of a list including:
\itemize{
\item a summary of the data used:
\item A summary including:
}
\itemize{
\item Nbirths: the raw number of births
\item Nadults: the raw number of potential reproductive adults
\item NOffsp: the number of offspring with known birth date and with parents with known age
\item NParent: the number of unique parents with known age and with offspring with known birth date
\item NOffsp_age: the number of offspring with known birth date and with parents with a known age > 0
\item NParent_age: the number of unique parents with known and positive age and with offspring with known birth date
\item a logical indicated if any reproductive analysis can be performed
\item If no analyses can be performed, an error and its number (Nerr) are returned: The possibility for  this functions are: 1/Nbirths < minNrep, 2/Nparents < minNparep, and 3/Data from N Institution
\item a logical indicated if the age at maturity analysis can be performed
\item a logical indicated if the litter size analysis can be performed
}
\itemize{
\item The output of each analysis. See the function Rep_agemat, Rep_littersize, Rep_fertility, and and Rep_seasonality for more details.
}
}
\description{
Run the age at sexual maturity analysis, the litter size analysis
}
\examples{
data(core)
data(collection)
data(parent)
data(moves)
out <- Rep_main (coresubset = core, collection, parent, moves,  
          Repsect = c('agemat', 'litter'),
          minNrep = 1, minNparep = 1,
          BirthType_parent = "Captive", BirthType_offspring = "Captive", 
         ) 
  
}
