% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_zimsdata.R
\name{Split_Zimsdata}
\alias{Split_Zimsdata}
\title{Split raw extracted data}
\usage{
Split_Zimsdata(ZIMSDir, ExtractDate = NULL, silent = FALSE)
}
\arguments{
\item{ZIMSDir}{\code{character} directory where to find data}

\item{ExtractDate}{\code{character} Extraction Date}

\item{silent}{\code{logical} Whether information of advancement should be printed}
}
\value{
Save all tables per taxa after:
\itemize{
\item selecting only "Individual" for Animal Type
\item adding binSpecies name
\item for weights and lengths: checking that measurement are > 0 and adding the following columns: Age in years, UnitOfMeasure and MeasurementValue.
}
}
\description{
Check Data and split tables per Taxa
}
\examples{

file = system.file("sci_Animal.csv", package = "ISRverse")
ZIMSDirtest = dirname(file)

Split_Zimsdata(ZIMSDir = ZIMSDirtest, ExtractDate = "2024-08-29")

list.files(ZIMSDirtest, recursive = TRUE)
}
