% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_90.R
\name{Sur_90}
\alias{Sur_90}
\title{Maximum Longevity}
\usage{
Sur_90(
  theMat,
  model = "GO",
  shape = "bathtub",
  ncpus = 1,
  ageMax = 120,
  dage = 0.01
)
}
\arguments{
\item{theMat}{\code{array} including the posteriors estimates of the model parameter}

\item{model}{\code{character} names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{ncpus}{\code{numeric} Number of core to use}

\item{ageMax}{\code{numeric} Maximum age in years Default = 120}

\item{dage}{\code{numeric} precision  for age Default = 0.01}
}
\value{
a data frame including age, the mean and 95\% credible interval of the L90
}
\description{
Estimate L90: age at which 90\% of the population died
}
\examples{
theMat = as.matrix(data.frame( b0 = rnorm(10, -6, 0.01),
                               b1= rnorm(10, 0.1, 0.01)))


out <- Sur_90(theMat, model = 'GO', shape = 'simple', ncpus = 2,
              ageMax = 50, dage = 0.1)
}
