% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_ana.R
\name{Sur_ana}
\alias{Sur_ana}
\title{Survival Analysis}
\usage{
Sur_ana(
  sexData,
  outlLev1 = 100,
  models = "GO",
  shape = "simple",
  mindate = "1980-01-01",
  minNsur = 50,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2
)
}
\arguments{
\item{sexData}{\code{data.frame} including at least the following columns \emph{anonID}, \emph{Birth.Date} (\code{date}), \emph{Depart.Date} (\code{date}), \emph{Entry.Date} (\code{date}), \emph{Max.Birth.Date} (\code{date}), \emph{Min.Birth.Date} (\code{date}), \emph{Entry.Type}, \emph{Depart.Type} and \emph{AnonInstitutionID}}

\item{outlLev1}{\code{numeric} Start threshold used to selected for the data: 100\%, 99.9, 99 or 95\%}

\item{models}{\code{vector of characters} names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{mindate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{minNsur}{\code{numeric} Minimum number of individual records needed to run the survival analysis. Default = 50}

\item{minlx}{\code{numeric} between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1}

\item{MinBirthKnown}{\code{numeric} between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3}

\item{niter}{\code{numeric}. number of MCMC iterations to run. see ?basta for more information. Default = 25000}

\item{burnin}{\code{numeric} Number of iterations removed so that the model has time to converge. see ?basta for more information. Default = 5001}

\item{thinning}{\code{numeric} Number of iteration to run before saving a set of parameters. see ?basta for more information. Default = 20}

\item{nchain}{\code{numeric} Number of chains to run. Default = 5001}

\item{ncpus}{\code{numeric} Number of computer core to use. Default = 2}
}
\value{
The output of a list including:
\itemize{
\item a summary of the data used:
}
\itemize{
\item NGlobal: Number of captive born individuals selected from global collections
\item NBasta: Number of data (individuals) selected for the BaSTA/survival analysis
\item Ndead:Number of individuals with known age of death used in the BaSTA/survival analysis
\item maxAge: Maximum age of known age individuals
\item maxAlive:Maximum number of years spent ex situ
\item lxMin:Minimum survivorship reached with the raw Kaplan-Meier model
\item OutLev: threshold selected for the distribution of  time spent alive: 100\%, 99.9, 99 or 95\%
\item a logical indicated if the growth analysis was performed
\item If the survival analysis was not performed, an error and its number (Nerr) are returned: The possibility for  this functions are: 1/No raw data and 2/lxMin > minlx 3/NBasta = 0 4/ \%known births < MinBirthKnown 5/Data from 1 Institution 6/Nbasta < minNsur, 7/no DIC from Basta.
}
\itemize{
\item the basta fit of the best model
\item the DIC table comparing the different fit of the models
}
}
\description{
Run basta models on the data after checking conditions for correct convergence of the models
}
\examples{
data(core)
out <- Sur_ana(core,  models = "GO", shape = "simple",
                niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)
}
