% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_out.R
\name{Sur_out}
\alias{Sur_out}
\title{Create output and plots for the survival analysis}
\usage{
Sur_out(
  out,
  shape = "simple",
  XMAX = 120,
  ncpus = 2,
  minAge = 0,
  PlotDir = NULL,
  plotname = "",
  Min_MLE = 0.1,
  MaxLE = 2,
  firstyear = FALSE
)
}
\arguments{
\item{out}{\code{numeric} The output of the survival analysis}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{XMAX}{\code{numeric} Maximum possible age. Default = 120}

\item{ncpus}{\code{numeric} Number of computer core to use. Default = 2}

\item{minAge}{\code{numeric} Ages at which the analyses should start.  see ?basta for more information. Default = 0}

\item{PlotDir}{\code{character} Directory to save the plots. Default = NULL, no plot is saved}

\item{plotname}{\code{character} Name used to save the plot. Default = ""}

\item{Min_MLE}{\code{numeric} Goodness of fit: Minimum survivorship at Mean life expectancy}

\item{MaxLE}{\code{numeric} Goodness of fit: Maximum remaining life expectancy at max age}

\item{firstyear}{\code{logical} Whether the analysis was done only on first year. Default = FALSE}
}
\value{
The output list of the survival analysis including in addition:
\itemize{
\item a summary of the data used:
}
\itemize{
\item A goodness of fit tested the trend in the residuals between the model prediction and the kaplan meier estimator
}
\itemize{
\item the estimated remaining life expectancy per age
\item the age at which 90\% of the population died
\item the age at which 50\% of the population died
\item the estimated probability to live one year more
\item the estimated probability to live five years more
If PlotDir is filled, 2 plots are produced: one showing the outliers removed from the data, and one showing the fit of the model on the data.
}
}
\description{
Estimate remaning life expectancy, L90, L50 and age-specific survival.
}
\examples{
data(core)
data(deathinformation)
out <- Sur_ana(core,  DeathInformation = deathinformation, models = "GO", shape = "simple",
               niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)

out <- Sur_out(out, shape = "simple", ncpus = 3)
}
