% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_relex.R
\name{Sur_relex}
\alias{Sur_relex}
\title{Remaining life expectancy}
\usage{
Sur_relex(Lx, dx = 0.01, xv)
}
\arguments{
\item{Lx}{\code{numeric} Survivorship}

\item{dx}{\code{numeric} Precision for age}

\item{xv}{\code{numeric} Age vector}
}
\value{
a data frame including age, the mean and 95\% credible interval of the remaining life expectancy
}
\description{
Estimate remaining life expectancy over ages.
}
\examples{
Lx = matrix(c(seq(1,0,by = -0.1), 1,seq(0.5,0,length.out = 10)),nrow =2)
out <- Sur_relex(Lx, dx = 1, xv = c(0:10))
}
