% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_bastab.R
\name{surv_Bastab}
\alias{surv_Bastab}
\title{Create Basta data frame}
\usage{
surv_Bastab(
  coresubset,
  DeathInformation,
  EarliestDate = NA,
  LatestDate = NA,
  OtherCovars = NA,
  ExcludeStillBirth = FALSE
)
}
\arguments{
\item{coresubset}{\code{data.frame} including at least the following columns \emph{AnimalAnonID}, \emph{binSpecies}, \emph{Class}, \emph{Order}, \emph{Family}, \emph{CommonName}, \emph{BirthDate}, \emph{DepartDate}, \emph{EntryDate}, \emph{MaxBirthDate}, \emph{MinBirthDate}, \emph{EntryType}, and \emph{DepartType}.}

\item{DeathInformation}{\code{data.frame} including at least the following columns \emph{AnimalAnonID} and \emph{RelevantDeathInformationType}.}

\item{EarliestDate}{\code{character 'YYYY-MM-DD'} Earlier date to be included.}

\item{LatestDate}{\code{character 'YYYY-MM-DD'} Latest date to be included.}

\item{OtherCovars}{\code{vector of character}. Additional variables to include in the data.}

\item{ExcludeStillBirth}{\code{logical} Whether to exclude still births.}
}
\value{
the subset dataset in the basta format
}
\description{
Check column names and succession of dates to prepare the data for BASTA.
}
\details{
This function removes:
\itemize{
\item individuals with NA in the columns BirthDate, MinBirthDate, MaxBirthDate, EntryDate, and DepartDate.
\item individuals for which the dates from min birth date/entry date to Depart date do not follow one another in a logical order.
\item Still born individuals if required
\item individuals with depart date anterior to earliest date
\item individuals with entry date posterior to latest date
\item Depart date posterior to latest date are changed to latest date. These individuals are considered as right-censored.
}
}
\examples{
data(core)
data(deathinformation)
out<- surv_Bastab(core, DeathInformation = deathinformation,
                  EarliestDate = '1990-01-01', LatestDate = '2020-12-31', 
                  OtherCovars = "SexType", ExcludeStillBirth = TRUE)


}
