% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_devout.R
\name{Tx_devout}
\alias{Tx_devout}
\title{Summarize taxon profiles analyses}
\usage{
Tx_devout(
  SpeciesTable,
  AnalysisDir,
  SaveDir,
  namefile = "",
  taxaList = "Mammalia",
  BySex = list(Mammalia = c("Male", "Female")),
  Sections = c("sur", "rep", "gro")
)
}
\arguments{
\item{SpeciesTable}{\code{Data.frame} including 2 columns: \emph{Class} and \emph{Species}}

\item{AnalysisDir}{\code{character} directory where to find the .Rdata files}

\item{SaveDir}{\code{character} directory where to save the json files}

\item{namefile}{\code{character} Suffix to add to the name of files produced if needed. Default = ""}

\item{taxaList}{\code{vector of characters} name of the taxa studied. Default= "Mammalia"}

\item{BySex}{\code{List} indicating the sexes analyzed. Default= list(Mammalia= c("Male", "Female"))}

\item{Sections}{\code{vector of character} names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")}
}
\value{
It saves summary tables and plots for each Sections and a general summary table. It returns the main summary table
}
\description{
Produce summary tables and plots of the demographic analyses made for all species
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
AnalysisDir  = paste0(dirname(file),'\\\\Rdata')
SaveDir = paste0(tempdir(check = TRUE),'\\\\temp')
dir.create(SaveDir)
SpeciesTable = data.frame( Class = "Reptilia", Species = "Testudo hermanni")
Tx_devout(SpeciesTable, AnalysisDir, SaveDir,
                        taxaList = "Reptilia",
                        BySex = list(Reptilia = c("Male", "Female")) ,
                        Sections = c("sur", 'gro')
)
list.files(SaveDir)
unlink(SaveDir, recursive = TRUE)
}
