% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_analysis.R
\name{Gro_analysis}
\alias{Gro_analysis}
\title{Growth model selection}
\usage{
Gro_analysis(
  data_weight,
  all_mods = c("vonBertalanffy"),
  percentiles = c(2.5, 97.5)
)
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the numeric columns \emph{Age}, \emph{MeasurementValue} and \emph{AnimalAnonID}}

\item{all_mods}{\code{vector of character} indicatingthe growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"}

\item{percentiles}{\code{vector of numeric} indicating the percentiles that need to be estimated. default = c(2.5,97.5) corresponding to the 95\% predicted interval.}
}
\value{
a list including:
\itemize{
\item a data frame with the percentile selected
\item the fit of the best model
\item the AIC table of the model
\item GOF: 3 GOF test including the normality of residuals and the test of trend in residuals against X and in the variance of the residuals
}
}
\description{
This function fit a serie of growth models to a dataset, select the best one by AIC and estimates the percentiles of the predicted distribution of values.
}
\examples{
Age <- sample(c(0:10), 100, replace = TRUE)
AnimalAnonID <- sample(c(0:20), 100, replace = TRUE)
MeasurementValue <- exp(0.2+15 * (1 - exp(-(0.1) * log(Age+1)))+ 
                          rnorm(100,0,0.01) + AnimalAnonID*0.1)-1 
dat = data.frame(Age = Age, MeasurementValue = MeasurementValue, 
                 AnimalAnonID = AnimalAnonID, MeasurementType = "Live Weight")

out = Gro_analysis(dat, 
                   all_mods = c("logistic", "vonBertalanffy", "gam"), 
                   percentiles = c(2.5, 97.5))
}
