% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_cleanmeasures.R
\name{Gro_cleanmeasures}
\alias{Gro_cleanmeasures}
\title{Prepare weight and length Dataset}
\usage{
Gro_cleanmeasures(
  data,
  coresubse,
  BirthType = "All",
  type = "weight",
  MeasureType = NULL,
  sex = "All",
  mindate = "1980-01-01",
  corevariable = NULL,
  variablekeep = NULL
)
}
\arguments{
\item{data}{\code{data.frame} including the following columns \emph{MeasurementValue}, \emph{MeasurementDate}, \emph{CollectionScopeType}, \emph{UnitOfMeasure}, \emph{ExcludedFromNorms}, \emph{EstimatedMeasurement}, \emph{RecordType}, \emph{MeasurementType}, \emph{anonID} and \emph{AnonInstitutionID}.}

\item{coresubse}{\code{data.frame} including at least the following columns \emph{binSpecies}, \emph{Sex}, \emph{anonID} and \emph{Birthdate} (\code{date})}

\item{BirthType}{\code{character} Captive, Wild, or All Default =  TRUE}

\item{type}{\code{character} Either 'weight' or 'length'. Default =  'weight}

\item{MeasureType}{\code{vector of characters} Name of the type of measurements that should be included. Default = NULL, all measurement type are included.}

\item{sex}{\code{character} Male, Female or All Default =  "All"}

\item{mindate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{corevariable}{\code{vector character} Name of the core columns that must be added to data}

\item{variablekeep}{\code{vector character} Name of the data columns that must also be returned}
}
\value{
A list including
\itemize{
\item The data frame including selected measures, plus individual birth date and individual age at each measure
\item A summary with the number of individuals and weight:
- NInd_raw  and, NWeight_raw indicate the number of unique individuals and the number of weights once the sex, the birth type, the measurment type from global collection have been selected
- NWeight_val and NInd_val indicate the number of unique individuals and the number of weights once valid measurment from Husbandry have been selected
- NWeight_age = 0 and NInd_age indicate the number of unique individuals and the number of weights once negative age has been removed.
-  If no data are selected, an error and its number (Nerr) are returned: The possibility for this functions are: 1/No raw data and 2/No valid weight measure.
}
}
\description{
Select clean measures, unify the unity used and add individual age at each measure
}
\examples{
data(raw_weights)
data(core)
data_weights= Gro_cleanmeasures(raw_weights, core,
                                BirthType = "All",
                                MeasureType = 'Live weight',
                                sex = "All", 
                                mindate = "1980-01-01")
}
