% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_remoutliers.R
\name{Gro_remoutliers}
\alias{Gro_remoutliers}
\title{Remove outliers from weigth or length data}
\usage{
Gro_remoutliers(
  data_weight,
  Taxa,
  AgeMat = NA,
  maxweight = NULL,
  variableid = "AnimalAnonID",
  min_Nmeasures = 7,
  perc_weight_min = 0.2,
  perc_weight_max = 2.5,
  IQR = 2.75,
  minq = 0.025,
  Ninterval_juv = 10
)
}
\arguments{
\item{data_weight}{\code{data.frame} including at least the following columns : \emph{MeasurementValue} (\code{numeric}), \emph{MeasurementType},  \emph{Age}  (\code{numeric}).}

\item{Taxa}{\code{character} Name of the taxa.}

\item{AgeMat}{\code{numeric} Age at sexual maturity to differentiate juveniles (still growing) from adults.}

\item{maxweight}{\code{numeric} the maximum weight allowed for the dataset}

\item{variableid}{\code{character} name of the variable including individual ids. It must also be a column of data_weight.}

\item{min_Nmeasures}{\code{integer} Minimum number of measures for an individual to check outliers along its growth trajectory. It must be at least 5.}

\item{perc_weight_min}{\code{numeric} Minimum percentage of weight that an individual can naturally lose or gain during a year. Default: 20\%.}

\item{perc_weight_max}{\code{numeric} Maximum percentage of weight that an individual can naturally lose or gain during a year. Default: 250\%.}

\item{IQR}{\code{numeric} influences the sensitivity of the function to remove outliers using log-linear model of growth. It should be above 1 Default =  2.75. A higher number makes the function less sensitive to find outliers.}

\item{minq}{\code{numeric} Sensitivity of the function to remove outliers using percentiles, between 0 and 1.}

\item{Ninterval_juv}{\code{integer} Number of intervals used for the sliding windows for juveniles.}
}
\value{
The data frame including the additional column \code{KEEP} a numeric vector of 1 and 0 indicating the measures to keep. The 0 signal outliers. Other additional columns keep1, keep2, keep3 indicates the individuals highlighted as outliers (0) in steps 1 to 3 and the column juv indicates which individuals have been classified as juveniles.
}
\description{
Takes a data frame including weight measures and Age and look for possible outliers
}
\details{
This function follow 4 different steps to highlight outliers:
1/ It removes the measures higher than \code{maxweight} (Instead of using the OrdMag Fernando that was removing only very extreme values, a maximum value was set for each taxa).
2/ Independently on adults and juveniles, it uses the function Gro_Rout_quan() to removes outliers based on percentiles on adults and juveniles. It uses sliding windows of age for juveniles.
3/  Independently on adults and juveniles, it uses the function Gro_lin_ind() to build generalized additive models for each individual trajectories with at least 7 measures and remove outliers based on the residuals of the models.
4/  It uses the function Gro_lin_ind() to build a common generalized additive model for all growth trajectories and remove outliers based on the residuals of the models.
}
\examples{
data(weights)
weights = Gro_remoutliers(weights[weights$MeasurementType == "Live weight",], 
                          Taxa = "Mammalia", AgeMat = 10)
}
