% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRverse-package.R
\docType{package}
\name{ISRverse-package}
\alias{ISRverse}
\alias{ISRverse-package}
\title{ISRverse: Run ISR Analyses For Taxon Profiles}
\description{
Survival, Reproduction and Growth analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fplard/ISRverse}
  \item Report bugs at \url{https://github.com/fplard/ISRverse/issues}
}

}
\author{
\strong{Maintainer}: Floriane Plard \email{floriane.c.plard@gmail.com} [contractor]

Other contributors:
\itemize{
  \item Species 360 \email{Support@Species360.org} [copyright holder]
  \item Fernando Colchero [contributor]
}

}
\keyword{internal}
