% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zimsdata.R
\name{Load_Zimsdata}
\alias{Load_Zimsdata}
\title{Load Zims data}
\usage{
Load_Zimsdata(
  Taxa,
  ZIMSDir,
  Species,
  Animal = FALSE,
  tables = c(),
  silent = FALSE
)
}
\arguments{
\item{Taxa}{\code{character} Names of the taxa.}

\item{ZIMSDir}{\code{character} Directory where to find data.}

\item{Species}{\code{list of character} names of the species studied. A list per taxa should be used, or "All".}

\item{Animal}{\code{logical} Whether the core animal data should be loaded.}

\item{tables}{\code{vector character} the type of data to load.}

\item{silent}{\code{logical} Whether information of advancement should be printed.}
}
\value{
A list of the data frames
}
\description{
Load ZIMS data frames of given taxa
}
\details{
\code{tables} can take the following values : "Contraception", "HealthStatus", "DeathInformation","Institution","InstitutionAssociation", "Move","Parent", "Weight", "Length","DeathInformation", "Collection"
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSDirtest = dirname(file)
Split_Zimsdata(ZIMSDir = ZIMSDirtest)

data <- Load_Zimsdata	(Taxa = "Reptilia", 
                       ZIMSDir = ZIMSDirtest, 
                       Species = list(Reptilia = "All"),
                       Animal = TRUE,
                       tables = 'Weight') 
data$Reptilia$Animal
unlink(glue::glue("{ZIMSDirtest}/Split_Reptilia"), recursive = FALSE)
}
