% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zimsdata.R
\name{Load_Zimsdata}
\alias{Load_Zimsdata}
\title{Load Zims data}
\usage{
Load_Zimsdata(
  taxa,
  ZIMSdir,
  species = "All",
  Animal = FALSE,
  tables = c(),
  silent = FALSE
)
}
\arguments{
\item{taxa}{\code{character} the name of the taxa studied}

\item{ZIMSdir}{\code{character} directory where to find data}

\item{species}{\code{list of character} names of the species studied. A list per taxa name should be used, or Default = "All"}

\item{Animal}{\code{logical} Whether the core animal data should be loaded Default = TRUE}

\item{tables}{\code{vector character} the type of data to load.}

\item{silent}{\code{logical} Whether information of advancement should be printed}
}
\value{
A list of the data frames
}
\description{
Check if selected data are already in the global environment. If not, it load them.
}
\details{
\code{type} can take the following values : "Contraception", "HealthStatus", "DeathInformation","Institution","InstitutionAssociation", "Move","Parent", "Weight", "Length","DeathInformation", "Collection"
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdirtest = dirname(file)
Split_Zimsdata	(ZIMSdir = ZIMSdirtest)

data <- Load_Zimsdata	(taxa = "Reptilia", 
                       ZIMSdir = ZIMSdirtest, 
                       Animal = TRUE,
                       tables = 'Weight') 
data$Reptilia$Animal
unlink(glue::glue("{ZIMSdirtest}/Split_Reptilia"), recursive = FALSE)
}
