% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{Summarize taxon profiles analyses}
\usage{
make_summary(
  AnalysisDir,
  SaveDir,
  namefile = "",
  taxaList = "Mammalia",
  BySex = list(Mammalia = c("Male", "Female")),
  Sections = c("sur", "rep", "gro")
)
}
\arguments{
\item{AnalysisDir}{\code{character} directory where to find the .Rdata files}

\item{SaveDir}{\code{character} directory where to save summary plots and tables}

\item{namefile}{\code{character} Suffix to add to the name of files produced if needed. Default = ""}

\item{taxaList}{\code{vector of character} names of the taxa studied. Default= "Mammalia"}

\item{BySex}{\code{list} of the taxa names indicating the sexes analyzed. Default=list(Mammalia = c("Male", "Female"))}

\item{Sections}{\code{vector of character} names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")}
}
\value{
It saves summary tables and plots for each Sections and a general summary table. It returns the main summary table
}
\description{
Produce summary tables and plots of the demographic analyses made for all species
}
\examples{
file = system.file("Testudo_hermanni.RData", package = 'ISRverse')
AnalysisDir  = dirname(file)
SaveDir = paste0(tempdir(check = TRUE),'\\\\temp')
dir.create(SaveDir)

SummTab <- make_summary(AnalysisDir, SaveDir,
               taxaList = "Reptilia",
               BySex = list(Reptilia = c("Male", "Female")) ,
               Sections = c("sur", 'gro')
)
list.files(SaveDir)


unlink(SaveDir, recursive = TRUE)
}
