% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_animal.R
\name{Prep_Animal}
\alias{Prep_Animal}
\title{Prepare Animal data}
\usage{
Prep_Animal(Animal, minDate = "1900-01-01", extractDate)
}
\arguments{
\item{Animal}{\code{data.frame} including at least the following columns \emph{AnimalAnonID},  \emph{BirthDate}, \emph{BirthDateEstimateType}, \emph{BirthDateEstimateStart}, \emph{BirthDateEstimateEnd}, \emph{BirthType}, \emph{FirstAcquisitionDate}, \emph{DeathDate}, \emph{DeathDateEstimateType}, \emph{DeathDateEstimateStart}, \emph{DeathDateEstimateEnd}, \emph{LastCommentEntryDate}, \emph{BirthObserved}, \emph{GlobalStatus}, and  \emph{LastTXDate}}

\item{minDate}{\code{character}: Earlier possible date: date used when minimum birth or death date unknown}

\item{extractDate}{\code{character}: Latest possible date in the data: date used when individuals are still alive}
}
\value{
The checked and prepared dataset
}
\description{
Check dates and add needed columns to animal data
}
\details{
This function add the columns: MinBirthDate, MaxBirthDate, MinDeathDate, MaxDeathDate, EntryDate, EntryType, DepartDate, DepartType, Birth_Uncertainty, and Death_Uncertainty. It removes individuals for which dates are not chronologically meaningful (i.e. birth date after death date) and for which Birth_Uncertainty is higher than \code{uncert_birth}.
}
\examples{
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdirtest = dirname(file)

data <- Load_Zimsdata	(taxa = "Reptilia",
                       ZIMSdir = ZIMSdirtest,
                       Animal = TRUE)

Animal <- Prep_Animal(data$Reptilia$Animal, extractDate = lubridate::as_date("2023/12/23"))

unlink(ZIMSdirtest, recursive = FALSE)
}
