% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_agemat.R
\name{Rep_agemat}
\alias{Rep_agemat}
\title{Age of reproduction}
\usage{
Rep_agemat(ReproData)
}
\arguments{
\item{ReproData}{\code{data frame} including at least the columns \emph{ParentAnonID} and \emph{Parent_Age}.}
}
\value{
A data frame including:
\itemize{
\item N_moth_agemat: Number of mothers used
\item N_birth_agemat : Number of births used
\item ageMat: age at first reproduction (i.e. when 2.5\% of females have reproduced for the first time)
\item ageYouRep: Minimum age at first reproduction
\item ageOld1Rep: Maximum age at first reproduction
\item ageMean1Rep: Mean age at first reproduction
\item ageSd1Rep: Standard deviation of age at first reproduction
\item ageOldRep: Maximum age of reproduction
\item ageMeanRep: Mean age of reproduction
\item ageSdRep: Standard deviation of age of reproduction
\item ageMedRep: Median age of reproduction
\item ageMeanRepIC: 95\% confident interval of mean age of reproduction
\item ageMean1RepIC: 95\% confident interval of mean age at first reproduction
}
}
\description{
Returns summary statistcs for age of reproduction
}
\examples{
data(core)
data(collection)
data(parent)
data(moves)
#prepare Data
Data <- Rep_prepdata (coresubset = core, collection, parent, moves, 
                      MinNRepro = 1, MinNPaRepro = 1)
#Calculate summary for reproductive ages
out <- Rep_agemat(Data$ReproData)
out

}
