% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_prepdata.R
\name{Rep_prepdata}
\alias{Rep_prepdata}
\title{Prepare Reproduction date}
\usage{
Rep_prepdata(
  coresubset,
  collection,
  parent,
  moves,
  BirthType_parent = "Captive",
  BirthType_offspring = "Captive",
  Age_uncert = 1e+06,
  Global = TRUE
)
}
\arguments{
\item{coresubset}{\code{data frame} cleaned core data including only the selected individuals. It must includes at least the following columns: \emph{anonID}, 'birthType', \emph{MaxBirthDate}, \emph{MinBirthDate}, \emph{DepartDate}, \emph{BirthDate},\emph{Sex}, and \emph{firstInst}}

\item{collection}{\code{data frame} Collection data including at least the following columns: \emph{AnimalID}, \emph{ScopeType}, \emph{ChangeDate}}

\item{parent}{\code{data frame} Parent data including at least the following columns: \emph{ParentAnonID}, \emph{ParentCollectionScopeType}, \emph{OffspringCollectionScopeType}, \emph{AnonID}, \emph{ParentOriginType}, \emph{Probability}}

\item{moves}{\code{data frame} Moves data including at least the following columns: \emph{AnonID}, \emph{To}, \emph{Date}}

\item{BirthType_parent}{\code{character} Captive, Wild, or All. Default =  "Captive"}

\item{BirthType_offspring}{\code{character} Captive, Wild, or All. Default =  "Captive"}

\item{Age_uncert}{\code{numeric} Maximum uncertainty (= MaxBirthDate - MinBirthDate) in birth date in days. Default = 10000}

\item{Global}{\code{logical} Whether only individuals belonging to global collections should be used. Default = TRUE}
}
\value{
A list including
\itemize{
\item The reproduction data
\item The potential reproductive Adults
\item A summary including:
}
\itemize{
\item Nbirths: the raw number of births
\item Nadults: the raw number of potential reproductive adults
\item NOffsp: the number of offspring with known birth date and with parents with known age
\item NParent: the number of unique parents with known age and with offspring with known birth date
\item NOffsp_age: the number of offspring with known birth date and with parents with a known age > 0
\item NParent_age: the number of unique parents with known and positive age and with offspring with known birth date
}
}
\description{
Prepare reproduction data including age of parent at birth, offspring sex and birth date and data of potential reproductive individuals at each age.
}
\examples{
data(core)
data(collection)
data(parent)
data(moves)

Data <- Rep_prepdata (coresubset = core, collection, parent, moves,
                      BirthType_parent = "Captive", BirthType_offspring = "Captive"
)
}
