% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_txprofile.R
\name{run_txprofile}
\alias{run_txprofile}
\title{Run Taxon Profiles}
\usage{
run_txprofile(
  taxa,
  Species_list,
  ZIMSdir,
  AnalysisDir,
  PlotDir,
  extractDate,
  minDate = "1980-01-01",
  Sections,
  erase_previous = FALSE,
  sexCats = c("Male", "Female"),
  minBirthDate = "1900-01-01",
  inparallel = FALSE,
  ipara = 1,
  npara = 1,
  minN = 50,
  maxOutl = 99,
  spOutLev = NULL,
  Global = TRUE,
  uncert_birth = 365,
  uncert_death = 365,
  uncert_date = 365,
  minNsur = 50,
  maxNsur = NULL,
  minInstitution = 2,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  XMAX = 120,
  Min_MLE = 0.1,
  MaxLE = 2,
  minAge = 0,
  firstyear = FALSE,
  models_sur = "GO",
  shape = "bathtub",
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  Repsect = c("agemat", "litter"),
  Birth_Type = "Captive",
  lastdead = FALSE,
  parentProb_Sire = 80,
  parentProb_Dam = 80,
  minNlitter = 20,
  Nday = 7,
  minNrepro = 100,
  minNparepro = 30,
  minNseas = 50,
  minNgro = 100,
  minNIgro = 50,
  MeasureType = "Live weight",
  models_gro = "vonBertalanffy"
)
}
\arguments{
\item{taxa}{\code{character} the name of the taxa studied}

\item{Species_list}{\code{vector of character} Species selected}

\item{ZIMSdir}{\code{character} directory where to find data}

\item{AnalysisDir}{\code{character} directory where to save results}

\item{PlotDir}{\code{character} Directory to save the plots. Default = NULL, no plot is saved}

\item{extractDate}{\code{character 'YYYY-MM-DD'} Date of data extraction}

\item{minDate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{Sections}{\code{vector of character} names of the sections to update in the taxon profile results: "sur", "rep" and/or "gro". Default = c("sur", "rep", "gro")}

\item{erase_previous}{\code{logical} whether the current result file should be deleted (before being replaced). Default = FALSE}

\item{sexCats}{\code{character} Male, Female or All Default =  "All"}

\item{minBirthDate}{\code{character}: Earlier possible date: date used when minimum birth or death date unknown}

\item{inparallel}{\code{logical} Whether this function is run in parallel on different computer. In other words: should the species list be divided? Default = FALSE}

\item{ipara}{\code{numeric} Id number of the computer used to select the species to run}

\item{npara}{\code{numeric} number of computers used in parallel}

\item{minN}{\code{numeric} Minimum number of individuals. Default = 50}

\item{maxOutl}{\code{numeric} Maximum threshold for the longevity distribution. Default = NULL}

\item{spOutLev}{\code{vector of character} List of species for which the Maximum threshold for the longevity distribution.XXXXXXXXXXXXXXXXXXXXXXXXX}

\item{Global}{\code{logical} Whether only individuals belonging to global collections should be used.}

\item{uncert_birth}{\code{numeric}: Maximum uncertainty accepted for birth dates, in days}

\item{uncert_death}{\code{numeric}: Maximum uncertainty accepted for death dates, in days. Default = 365}

\item{uncert_date}{\code{numeric}: Maximum uncertainty accepted for measurement dates: weight, in days. Default = 365}

\item{minNsur}{\code{numeric} Minimum number of individuals to run the survival analysis. Default = 50}

\item{maxNsur}{\code{numeric} Maximum number of individual records to run the survival analysis. Default = NULL}

\item{minInstitution}{\code{numeric} Minimum number of institutions that should hold records to run the analyses. Default = 2}

\item{minlx}{\code{numeric} between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1}

\item{MinBirthKnown}{\code{numeric} between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3}

\item{XMAX}{\code{numeric} Maximum possible age. Default = 120}

\item{Min_MLE}{\code{numeric} Goodness of fit: Minimum survivorship at Mean life expectancy}

\item{MaxLE}{\code{numeric} Goodness of fit: Maximum remaining life expectancy at max age}

\item{minAge}{\code{numeric} Ages at which the analyses should start.  see ?basta for more information. Default = 0}

\item{firstyear}{\code{logical} Whether to do the analysis only on first year. Default = FALSE}

\item{models_sur}{\code{vector of characters} names of the survival basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the survival basta model to run: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{niter}{\code{numeric}. number of MCMC iterations to run the survival model. see ?basta for more information. Default = 25000}

\item{burnin}{\code{numeric} Number of iterations removed so that the survival model has time to converge. see ?basta for more information. Default = 5001}

\item{thinning}{\code{numeric} Number of iteration to run the survival model before saving a set of parameters. see ?basta for more information. Default = 20}

\item{nchain}{\code{numeric} Number of chains to run the survival model. Default = 5001}

\item{ncpus}{\code{numeric} Number of computer core to use. Default = 2}

\item{Repsect}{\code{character} names of the reproductive analyses to run: "agemat", "litter" and/or ...}

\item{Birth_Type}{\code{character} Captive, Wild, or All. Default =  "Captive"}

\item{lastdead}{\code{logical} Whether the longest lived individuals should be considered dead. Default = FALSE}

\item{parentProb_Sire}{\code{numeric} Minimum percentage of parentage probability to include for Sire. Default = 80}

\item{parentProb_Dam}{\code{numeric} Minimum percentage of parentage probability to include for Dam. Default = 80}

\item{minNlitter}{\code{numeric} Minimum number of litters to run the analysis. The data frame for litter size will be produced in all cases. Default = 30}

\item{Nday}{\code{numeric} Number of consecutive days over which the birth dates of a litter/clutch can be spread. Default = 7}

\item{minNrepro}{\code{numeric} Minimum number of birth records needed to run reproductive analyses. Default = 50}

\item{minNparepro}{\code{numeric} Minimum number of unique parent records needed to run reproductive analyses. Default = 30}

\item{minNseas}{\code{numeric} XXXXXXXXXXX}

\item{minNgro}{\code{numeric} Minimum number of weight needed to fit the growth models}

\item{minNIgro}{\code{numeric} Minimum number of unique individuals needed to fit the growth models}

\item{MeasureType}{\code{vector of characters} Name of the type of measurements that should be included.  Default = NULL, all measurement type are included.}

\item{models_gro}{\code{vector of characters} indicating the growth models that need to be fit.The following models are supported : logistic, gompertz, chapmanRichards, vonBertalanffy, polynomial. default = "vonBertalanffy"}
}
\value{
Save and replace the result file for each specie sin the list. The file is saved in {analysisDir}\\Rdata\\{taxa}_{species}.Rdata
}
\description{
Load data needed and run or update demographic analyses for the list of species selected
}
\examples{
# Here is an example but use directly your own ZIMSdir (directory to find data) 
# and analysisDir (directory to save analysis)
file = system.file("sci_Animal.csv", package = 'ISRverse')
ZIMSdir = dirname(file)
AnalysisDir = paste0(tempdir(check = TRUE),'\\\\temp')
PlotDir = paste0(AnalysisDir,'\\\\plot')
# dir.create(AnalysisDir)
dir.create(paste0(AnalysisDir,'\\\\Rdata'), recursive = TRUE)
dir.create(PlotDir)

# #This code run the survival analysis
# run_txprofile(taxa = "Reptilia", Species_list = "All",
#               ZIMSdir = ZIMSdir, AnalysisDir = AnalysisDir,
#               PlotDir = PlotDir,
#               extractDate = "",
#               minDate = "1980-01-01",
#               Sections = c("sur"),
#               sexCats = c('Male', 'Female'),
#               niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3
# )

list.files(PlotDir)
list.files(paste0(AnalysisDir,'\\\\Rdata'))

unlink(AnalysisDir, recursive = TRUE)
unlink(PlotDir, recursive = TRUE)
}
