% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_species.R
\name{select_species}
\alias{select_species}
\title{Extract data}
\usage{
select_species(
  speciesname,
  coresubset,
  collection,
  minDate,
  extractDate,
  Global = TRUE
)
}
\arguments{
\item{speciesname}{\code{character} latin name of the species selected}

\item{coresubset}{\code{data.frame} including at least the following columns \emph{anonID}, \emph{BirthDate} (\code{date}), \emph{DepartDate} (\code{date}), \emph{EntryDate} (\code{date}), \emph{MaxBirthDate} (\code{date}), \emph{MinBirthDate} (\code{date}), \emph{EntryType}, \emph{DepartType}, \emph{DepartFrom}, \emph{firstInst}, \emph{LastTXDate}, \emph{DeathDate}, \emph{lastInst},\emph{globStat}, \emph{LastCollectionScopeType}, and \emph{FirstCollectionScopeType}}

\item{collection}{\code{data.frame} including at least the following columns\emph{RecordingInstitution}, \emph{ChangeDate}, \emph{ScopeType}, and \emph{AnimalID}.}

\item{minDate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{extractDate}{\code{character 'YYYY-MM-DD'} Date of data extraction}

\item{Global}{\code{logical} Whether only individuals belonging to global collections should be used.}
}
\value{
The output of a list including:
\itemize{
\item a summary of the data used:
}
\itemize{
\item Nraw : Raw number of individuals selected from global collections
\item Ndate : Number of individuals with an entry date posterior to the minimum date
\item Nglobal: Number of individuals selected from global collections
\item Nalive: Number of individuals still alive
\item firstDate: Date of first record
\item maxAgeraw: Maximum observed age
}
\itemize{
\item The subseted dataset
}
}
\description{
Extract data of the specified species that fill the condition on minimum date and global collections
}
\examples{
data(core)
data(collection)
out<- select_species(speciesname = "Gorilla gorilla", coresubset = core, collection,
                     minDate = "1980-01-01", extractDate = "2023-01-01")
out$summary
out$data
}
