% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_zimsdata.R
\name{Split_Zimsdata}
\alias{Split_Zimsdata}
\title{Split raw extracted data}
\usage{
Split_Zimsdata(ZIMSdir, extractDate = NULL, silent = FALSE)
}
\arguments{
\item{ZIMSdir}{\code{character} directory where to find data}

\item{extractDate}{\code{character} Extraction Date}

\item{silent}{\code{logical} Whether information of advancement should be printed}
}
\value{
A list of the data frames
}
\description{
Check Data and split tables per taxa
}
\examples{

file = system.file("sci_Animal.csv", package = "ISRverse")
ZIMSdirtest = dirname(file)

Split_Zimsdata(ZIMSdir = ZIMSdirtest)

list.files(ZIMSdirtest, recursive = TRUE)
}
