% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_ana.R
\name{Sur_ana}
\alias{Sur_ana}
\title{Survival Analysis}
\usage{
Sur_ana(
  sexData,
  DeathInformation,
  outlLev1 = 100,
  models = "GO",
  shape = "simple",
  minAge = 0,
  mindate = "1980-01-01",
  minNsur = 50,
  maxNsur = NULL,
  uncert_death = 365,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  minInstitution = 1,
  lastdead = FALSE,
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2
)
}
\arguments{
\item{sexData}{\code{data.frame} including at least the following columns \emph{AnimalAnonID}, \emph{BirthDate}, \emph{DepartDate}, \emph{EntryDate}, \emph{MaxBirthDate}, \emph{MinBirthDate}, \emph{EntryType}, \emph{DepartType}, \emph{FirstHoldingInstitution}, \emph{LastHoldingInstitution}, \emph{SexType}, and \emph{BirthType}}

\item{DeathInformation}{\code{data.frame} including at least the following columns \emph{AnimalAnonID} and \emph{RelevantDeathInformationType}}

\item{outlLev1}{\code{numeric} Start threshold used to selected for the data: 100\%, 99.9, 99 or 95\%}

\item{models}{\code{vector of characters} names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{minAge}{\code{numeric} Age at which the analysis should start.  see ?basta for more information. Default = 0}

\item{mindate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{minNsur}{\code{numeric} Minimum number of individual records needed to run the survival analysis. Default = 50}

\item{maxNsur}{\code{numeric} Maximum number of individual records to run the survival analysis. Default = NULL}

\item{uncert_death}{\code{numeric}: Maximum uncertainty accepted for death date, in days}

\item{minlx}{\code{numeric} between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1}

\item{MinBirthKnown}{\code{numeric} between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3}

\item{minInstitution}{\code{numeric} Minimum number of institutions that should hold records to run the survival analysis. Default = 1}

\item{lastdead}{\code{logical} Whether the longest lived individuals should be considered dead. Default = FALSE}

\item{niter}{\code{numeric}. number of MCMC iterations to run. see ?basta for more information. Default = 25000}

\item{burnin}{\code{numeric} Number of iterations removed so that the model has time to converge. see ?basta for more information. Default = 5001}

\item{thinning}{\code{numeric} Number of iteration to run before saving a set of parameters. see ?basta for more information. Default = 20}

\item{nchain}{\code{numeric} Number of chains to run. Default = 5001}

\item{ncpus}{\code{numeric} Number of computer core to use. Default = 2}
}
\value{
The output of a list including:
\itemize{
\item a summary of the data used:
}
\itemize{
\item NSelect: Number of individuals selected from filters
\item NUncertdeath: Number of individuals selected after filter uncertainty in death
\item NBasta: Number of data (individuals) selected for the BaSTA/survival analysis
\item Ndead:Number of individuals with known age of death used in the BaSTA/survival analysis
\item maxAge: Maximum age of known age individuals
\item maxAlive:Maximum number of years spent ex situ
\item lxMin:Minimum survivorship reached with the raw Kaplan-Meier model
\item OutLev: threshold selected for the distribution of  time spent alive: 100\%, 99.9, 99 or 95\%
\item a logical indicated if the growth analysis was performed
\item If the survival analysis was not performed, an error and its number (Nerr) are returned: The possibility for  this functions are: 2/Nuncertdeath < minNsur; 3/lxMin > minlx; 4/NBasta = 0; 5/ \%known births < MinBirthKnown; 6/Data from 1 Institution; 7/Nbasta < minNsur; 8/Nbasta < maxNsur; 9/no DIC from Basta.
}
\itemize{
\item the basta fit of the best model
\item the DIC table comparing the different fit of the models
}
}
\description{
Run basta models on the data after checking conditions for correct convergence of the models
}
\examples{
data(core)
data(deathinformation)
out <- Sur_ana(core,  DeathInformation = deathinformation, models = "GO", shape = "simple",
               niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)
}
