% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_main.R
\name{Sur_main}
\alias{Sur_main}
\title{Main survival analysis for taxon profiles}
\usage{
Sur_main(
  data.core,
  DeathInformation,
  Birth_Type = "All",
  models = "GO",
  shape = "simple",
  firstyear = FALSE,
  outlLev1 = NA,
  minAge = 0,
  mindate = "1980-01-01",
  minNsur = 50,
  maxNsur = NULL,
  minInstitution = 2,
  uncert_death = 365,
  minlx = 0.1,
  MinBirthKnown = 0.3,
  Min_MLE = 0.1,
  MaxLE = 2,
  XMAX = 120,
  lastdead = FALSE,
  niter = 25000,
  burnin = 5001,
  thinning = 20,
  nchain = 3,
  ncpus = 2,
  PlotDir = NULL,
  plotname = ""
)
}
\arguments{
\item{data.core}{\code{data.frame} including at least the following columns \emph{AnimalAnonID}, \emph{BirthDate}, \emph{DepartDate}, \emph{EntryDate}, \emph{MaxBirthDate}, \emph{MinBirthDate}, \emph{EntryType}, \emph{DepartType}, \emph{FirstHoldingInstitution}, \emph{LastHoldingInstitution}, \emph{SexType}, and \emph{BirthType}}

\item{DeathInformation}{\code{data.frame} including at least the following columns \emph{AnimalAnonID} and \emph{RelevantDeathInformationType}}

\item{Birth_Type}{\code{character} When separated the analysis by sex category, the birth type to be selected: Captive, Wild, or All Default =  "All"}

\item{models}{\code{vector of characters} names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{firstyear}{\code{logical} Whether to do the analysis only on first year. Default = FALSE}

\item{outlLev1}{\code{numeric} Start threshold used to selected for the data: 100\%, 99.9, 99 or 95\%}

\item{minAge}{\code{numeric} Ages at which the analyses should start.  see ?basta for more information. Default = 0}

\item{mindate}{\code{character 'YYYY-MM-DD'} Earlier date to include data}

\item{minNsur}{\code{numeric} Minimum number of individual records needed to run the survival analysis. Default = 50}

\item{maxNsur}{\code{numeric} Maximum number of individual records to run the survival analysis. Default = NULL}

\item{minInstitution}{\code{numeric} Minimum number of institutions that should hold records to run the survival analysis. Default = 2}

\item{uncert_death}{\code{numeric}: Maximum uncertainty accepted for death date, in days}

\item{minlx}{\code{numeric} between 0 and 1. Minimum reached survivorship from the raw Kaplan Meier analysis needed to run the survival analysis. Default = 0.1}

\item{MinBirthKnown}{\code{numeric} between 0 and 1. Minimum proportion of individuals with a known birth month in the data. Default = 0.3}

\item{Min_MLE}{\code{numeric} Goodness of fit: Minimum survivorship at Mean life expectancy}

\item{MaxLE}{\code{numeric} Goodness of fit: Maximum remaining life expectancy at max age}

\item{XMAX}{\code{numeric} Maximum possible age. Default = 120}

\item{lastdead}{\code{logical} Whether the longest lived individuals should be considered dead. Default = FALSE}

\item{niter}{\code{numeric}. number of MCMC iterations to run. see ?basta for more information. Default = 25000}

\item{burnin}{\code{numeric} Number of iterations removed so that the model has time to converge. see ?basta for more information. Default = 5001}

\item{thinning}{\code{numeric} Number of iteration to run before saving a set of parameters. see ?basta for more information. Default = 20}

\item{nchain}{\code{numeric} Number of chains to run. Default = 5001}

\item{ncpus}{\code{numeric} Number of computer core to use. Default = 2}

\item{PlotDir}{\code{character} Directory to save the plots. Default = NULL, no plot is saved}

\item{plotname}{\code{character} Name used to save the plot. Default = ""}
}
\value{
The output of a list including per sex categories:
\itemize{
\item a summary of the data used:
}
\itemize{
\item NGlobal: Number of captive born individuals selected from global collections
\item NBasta: Number of data (individuals) selected for the BaSTA/survival analysis
\item Ndead:Number of individuals with known age of death used in the BaSTA/survival analysis
\item maxAge: Maximum age of known age individuals
\item maxAlive:Maximum number of years spent ex situ
\item lxMin:Minimum survivorship reached with the raw Kaplan-Meier model
\item OutLev: threshold selected for the distribution of  time spent alive: 100\%, 99.9, 99 or 95\%
\item a logical indicated if the survival analysis was performed
\item If the survival analysis was not performed, an error and its number (Nerr) are returned: The possibility for  this functions are: 2/Nuncertdeath < minNsur; 3/NBasta = 0; 4/ \%known births < MinBirthKnown; 5/Data from 1 Institution; 6/Nbasta < minNsur; 7/Nbasta > maxNsur; 8/no DIC from Basta; 9/Kaplan-Meier does not fit; 10/Min(Life_exp) >= MaxLE; 11/"lx_MLE < Min_MLE"; 12/"lxmin > minlx"; 13/Kaplan-Meier does not fit
\item A goodness of fit tested the trend in the residuals between the model prediction and the kaplan meier estimator
}
\itemize{
\item the basta fit of the best model
\item the DIC table comparing the different fit of the models
\item the estimated remaining life expectancy per age
\item the age at which 90\% of the population died
\item the estimated probability to live one year more
\item the estimated probability to live five years more
If PlotDir is filled, 2 plots are produced: one showing the outliers removed from the data, and one showing the fit of the model on the data.
}
}
\description{
Run the survival analysis per sex and per birth type
}
\examples{
data(core)
data(deathinformation)
out <- Sur_main(core, DeathInformation = deathinformation, Birth_Type = "All",
                models = "GO", shape = "simple",
                niter = 1000, burnin = 101, thinning = 10, nchain = 3, ncpus = 3)
}
