% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_relex.R
\name{Sur_relex}
\alias{Sur_relex}
\title{Remaining life expectancy}
\usage{
Sur_relex(
  theMat,
  model = "GO",
  shape = "bathtub",
  ncpus = 1,
  xMax = 120,
  dx = 0.01
)
}
\arguments{
\item{theMat}{\code{array} including the posteriors estimates of the model parameter}

\item{model}{\code{character} names of the basta models to run: "G0", "EX", "LO" and/or "WE". see ?basta for more information. Default = "GO"}

\item{shape}{\code{character} shape of the basta model: "simple", "Makeham", "bathtub".  see ?basta for more information. Default = "simple"}

\item{ncpus}{\code{numeric} Number of core to use}

\item{xMax}{\code{numeric} Maximum age in years Default = 120}

\item{dx}{\code{numeric} precision for age Default = 0.01}
}
\value{
a data frame including age, the mean and 95\% credible interval of the remaining life expectancy
}
\description{
Estimate remaining life expectancy over age from the parameters outputs of a Basta model.
}
\examples{
theMat = as.matrix(data.frame( b0 = rnorm(10, -6, 0.01),
                               b1= rnorm(10, 0.1, 0.01)))


out <- Sur_relex(theMat, model = 'GO', shape = 'simple', ncpus = 2,
                 xMax = 50, dx = 0.1)
}
